% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{compare_sex}
\alias{compare_sex}
\title{Compares inferred sex with specified sex}
\usage{
compare_sex(dnam, specified_sex, kmeans = kmeans_default,
  centroids = centroids_default, plot = T)
}
\arguments{
\item{dnam}{list-output of \link{preprocess}}

\item{specified_sex}{named vector with prespecified sex}

\item{kmeans}{passed on to \link{infer_sex}}

\item{centroids}{passed on to \link{infer_sex}}

\item{plot}{\code{TRUE} shows scatter plot with specified sex}
}
\value{
Returns confusion matrix showing accordance of inferred and specified sex
}
\description{
Compares inferred sex with specified sex
}
\seealso{
\link{preprocess}, \link{infer_sex}
}
